[FP-02] Added logging and observability features

+++ services/payment-service/internal/dto/PaymentRequestDTO.go
@@ -2,8 +2,8 @@ package dto
 
 // PaymentRequest defines the incoming JSON structure
 type PaymentRequestDTO struct {
-	SenderID   string  `json:"sender_id"`
-	ReceiverID string  `json:"receiver_id"`
-	Amount     float64 `json:"amount"`
-	Currency   string  `json:"currency"`
+	SenderID   string `json:"sender_id"`
+	ReceiverID string `json:"receiver_id"`
+	Amount     int64  `json:"amount"` // minor units, e.g. paise
+	Currency   string `json:"currency"`
 }
